- Start refactoring map upload to better support client uploads
