+++ requirements.txt
@@ -10,3 +10,5 @@ python-magic>=0.4.27
 pydantic>=2.10,<3.0
 gunicorn>=22.0.0,<23.0.0
 django-filter==25.1
+orjson>=3.10.15,<4.0
+structlog>=25.1.0,<26.0.0
